+++ src/rust_backend.rs
@@ -1,4 +1,4 @@
-use crate::GcdResult;
+use crate::{get_mod, GcdResult};
 use glass_pumpkin::{prime, safe_prime};
 use num_bigint::{BigInt, Sign, ToBigInt};
 use num_integer::Integer;
@@ -14,6 +14,7 @@ use serde::{
 use std::{
     cmp::{Eq, Ord, Ordering, PartialEq, PartialOrd},
     fmt::{self, Debug, Display},
+    iter::{Product, Sum},
     mem::swap,
     ops::{
         Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign, Shl, Shr, Sub,
@@ -39,6 +40,7 @@ from_impl!(|d: i64| BigInt::from(d), i64);
 from_impl!(|d: i32| BigInt::from(d), i32);
 from_impl!(|d: i16| BigInt::from(d), i16);
 from_impl!(|d: i8| BigInt::from(d), i8);
+iter_impl!();
 ord_impl!();
 serdes_impl!(|b: &Bn| b.0.to_str_radix(16), |s: &str| {
     BigInt::from_str_radix(s, 16)
@@ -57,13 +59,10 @@ impl Bn {
     /// Returns `(self ^ exponent) mod n`
     /// Note that this rounds down
     /// which makes a difference when given a negative `self` or `n`.
-    /// The result will be in the interval `[0, n)` for `n > 0`,
+    /// The result will be in the interval `[0, n)` for `n > 0`
     pub fn modpow(&self, exponent: &Self, n: &Self) -> Self {
-        let nn = if n.0 < BigInt::zero() {
-            -n.clone()
-        } else {
-            n.clone()
-        };
+        assert_ne!(n.0, BigInt::zero());
+        let nn = get_mod(&n);
         if exponent.0 < BigInt::zero() {
             match self.invert(&nn) {
                 None => Self::zero(),
@@ -79,37 +78,54 @@ impl Bn {
 
     /// Compute (self + rhs) mod n
     pub fn modadd(&self, rhs: &Self, n: &Self) -> Self {
-        let mut t = (self + rhs) % n;
+        let nn = get_mod(&n);
+        let mut t = (self + rhs) % &nn;
         if t < Bn::zero() {
-            t += n;
+            t += &nn;
         }
         t
     }
 
     /// Compute (self - rhs) mod n
     pub fn modsub(&self, rhs: &Self, n: &Self) -> Self {
-        let mut t = (self - rhs) % n;
+        let nn = get_mod(&n);
+        let mut t = (self - rhs) % &nn;
         if t < Bn::zero() {
-            t += n;
+            t += &nn;
         }
         t
     }
 
     /// Compute (self * rhs) mod n
     pub fn modmul(&self, rhs: &Self, n: &Self) -> Self {
-        let mut t = (self * rhs) % n;
+        let nn = get_mod(&n);
+        let mut t = (self * rhs) % &nn;
         if t < Bn::zero() {
-            t += n;
+            t += &nn;
         }
         t
     }
 
     /// Compute (self * 1/rhs) mod n
     pub fn moddiv(&self, rhs: &Self, n: &Self) -> Self {
-        let mut t = (self * rhs.invert(n).unwrap()) % n;
-        if t < Bn::zero() {
-            t += n;
+        let nn = get_mod(&n);
+        match rhs.invert(&nn) {
+            None => Self::zero(),
+            Some(r) => {
+                let mut t = (self * r) % &nn;
+                if t < Bn::zero() {
+                    t += &nn;
+                }
+                t
+            }
         }
+    }
+
+    /// Compute -self mod n
+    pub fn modneg(&self, n: &Self) -> Self {
+        let mut t = self.clone() % n.clone();
+        t = n.clone() - t.clone();
+        t %= n.clone();
         t
     }
 
@@ -262,19 +278,19 @@ impl Bn {
         }
     }
 
-    /// Generate a safe prime
+    /// Generate a safe prime with `size` bits
     pub fn safe_prime(size: usize) -> Self {
         let p = safe_prime::new(size).unwrap();
         Self(p.to_bigint().unwrap())
     }
 
-    /// Generate a prime
+    /// Generate a prime with `size` bits
     pub fn prime(size: usize) -> Self {
         let p = prime::new(size).unwrap();
         Self(p.to_bigint().unwrap())
     }
 
-    /// True if self is a prime number
+    /// True if a prime number
     pub fn is_prime(&self) -> bool {
         match self.0.to_biguint() {
             None => false,